UI: extract notification about detached file into separate provider
